btrfs-progs: add key derivation helper functions
One of the way to generate a key is to derive it using
the user provided passphrase. The helper function here
uses libscrypt to generate the key using the user provided
passphrase. Further it generates the key description
or key tag using the sha512.

The planned fscrypt cli tool isn't ready yet, when thats
ready, probably we shall use the lib tools from that.

Signed-off-by: Anand Jain <anand.jain@oracle.com>